ENGR00217401-3 gralloc: fix type error of phy memory address.
fix the phy memory adderss type, should be u32 rather that signed.

+++ mx6/libgralloc_wrapper/gralloc_priv.h
@@ -93,7 +93,7 @@ struct private_handle_t {
     int     size;
     int     offset;
     int     base;
-    int     phys;
+    unsigned long phys;
     int     format;
     int     width;
     int     height;
